build: Move headers as they are installed
