+++ src/cpp/include/firebolt/advertising.h
