+++ src/cpp/include/firebolt/lifecycle.h
