+++ src/cpp/include/firebolt/version.h.in
