+++ src/cpp/src/display_impl.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "display.h"
+#include "firebolt/display.h"
 #include <firebolt/helpers.h>
 
 namespace Firebolt::Display
