+++ src/cpp/src/json_types/jsondata_advertising_types.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "advertising.h"
+#include "firebolt/advertising.h"
 #include <firebolt/json_types.h>
 #include <nlohmann/json.hpp>
 #include <string>
