+++ src/cpp/src/json_types/jsondata_stats_types.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "stats.h"
+#include "firebolt/stats.h"
 #include <firebolt/json_types.h>
 #include <nlohmann/json.hpp>
