+++ src/cpp/test/component/deviceTest.cpp
@@ -17,16 +17,11 @@
  * limitations under the License.
  */
 
-#include "./component/utils.h"
-#include "firebolt.h"
+#include "firebolt/firebolt.h"
 #include "json_engine.h"
 #include "json_types/jsondata_device_types.h"
-#include "gtest/gtest.h"
-
-#include <condition_variable>
+#include <gtest/gtest.h>
 #include <iostream>
-#include <mutex>
-#include <thread>
 
 class DeviceTest : public ::testing::Test
 {
